double quoting (?)

+++ eslint.config.mjs
@@ -60,8 +60,8 @@ export default [
             'yml/file-extension': ['error', { extension: 'yaml', caseSensitive: true }],
             // Enforce 2-space indentation
             'yml/indent': ['error', 2],
-            // Enforce quote style: prefer single quotes, only when necessary
-            'yml/quotes': ['error', { prefer: 'single', avoidEscape: true }],
+            // Enforce quote style: prefer double quotes, only when necessary
+            'yml/quotes': ['error', { prefer: 'double', avoidEscape: true }],
             // Enforce plain scalars (unquoted) when possible - quotes only when necessary
             'yml/plain-scalar': ['error', 'always'],
             // Keep important rules
